fix(ios): pass page sizes as the Int32 the generated queries take

+++ ios/Shiori/Features/Library/LibraryAPI.swift
@@ -19,7 +19,7 @@ enum LibraryAPI {
     ) async throws -> LibraryPageResult {
         let query = ShioriGraphQL.LibraryPageQuery(
             status: GraphQLHelpers.graphQLNullable(status.map(Self.graphQLStatus)),
-            limit: .some(limit),
+            limit: .some(Int32(limit)),
             after: GraphQLHelpers.graphQLNullable(after)
         )
         let data = try await GraphQLHelpers.fetch(GraphQLClient.shared.apollo, query: query)
